+++ modules/jumpstart_ui/tests/src/Unit/Plugin/Block/PageHeadingBlockTest.php
@@ -62,7 +62,7 @@ class PageHeadingBlockTest extends UnitTestCase {
       ->getMock();
     $container->set('cache_contexts_manager', $this->cacheContextsManager);
 
-    $this->account = $this->prophesize(AccountInterface::class);
+    $this->account = $this->createMock(AccountInterface::class);
 
     $this->block = new PageHeadingBlock([], 'jumpstart_ui_page_heading', ['provider' => 'jumpstart_ui']);
     \Drupal::setContainer($container);
@@ -73,7 +73,7 @@ class PageHeadingBlockTest extends UnitTestCase {
    */
   public function testAccess() {
     $this->cacheContextsManager->method('assertValidTokens')->willReturn(TRUE);
-    $this->assertTrue($this->block->access($this->account->reveal()));
+    $this->assertTrue($this->block->access($this->account));
   }
 
   /**