+++ modules/jumpstart_ui/tests/src/Unit/Plugin/paragraphs/Behavior/HeroPatternBehaviorTest.php
@@ -45,6 +45,7 @@ class HeroPatternBehaviorTest extends UnitTestCase {
     $display_ids = [];
 
     $query = $this->createMock(QueryInterface::class);
+    $query->method('accessCheck')->willReturnSelf();
     $query->method('condition')->willReturnSelf();
     $query->method('execute')->willReturnReference($display_ids);