+++ modules/stanford_layout_paragraphs/src/Layouts/ThreeColumn.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Drupal\stanford_layout_paragraphs\Layouts;
+
+use Drupal\Core\Layout\LayoutDefault;
+
+/**
+ * Three column layout class.
+ */
+class ThreeColumn extends LayoutDefault {
+
+}